Window.onload event not anymore requires WebView::rendering (#2031)

+++ src/core/modules/threading/ParallelJobExecutor.h
@@ -64,7 +64,7 @@ public:
         return m_paramVector[index];
     }
 
-    int numberOfThread()
+    size_t numberOfThread()
     {
         return m_threadVector.size();
     }
